feat(Lottos): 로또 번호를 객체로 만들어 관리하는 Lotto 구현
